Rendered schematic.
The PNG is opaque white background.
The SVG is 'natural' (transparent background).

The 'plot' function of kicad generated the SVG.

The PNG was generated in Inkscape, from the SVG,
and required finding the background color and
changing it to white, and non-transparent.

Bothered to bring it to 192 dpi (96 dpi was
the default).

	new file:   kicad_22_dec_2021-a.png
	new file:   kicad_22_dec_2021-a.svg

On branch dvlp-aa